# Conflicts:
#	.github/workflows/linux_build.yml
#	.github/workflows/macOS_build.yml
#	.github/workflows/pr_cache_cleanup.yml
#	.github/workflows/windows_build.yml
#	CHANGELOG.md
#	README.md
#	appveyor.yml
#	ci-scripts/linux/tahoma-buildpkg.sh
#	ci-scripts/osx/tahoma-buildpkg.sh
#	toonz/cmake/BundleInfo.plist.in
#	toonz/installer/linux/deb-creator/deb-template/DEBIAN/control
#	toonz/installer/windows/setup.iss
#	toonz/sources/toonz/CMakeLists.txt
#	toonz/sources/toonz/cellselectioncommand.cpp
#	toonz/sources/xdg-data/CMakeLists.txt
#	toonz/sources/xdg-data/org.tahoma2d.Tahoma2D.metainfo.xml
#	ztoryc_generated/include/tversion.h

+++ toonz/sources/include/tools/toolutils.h
@@ -101,7 +101,7 @@ void drawRectWhitArrow(const TPointD &pos, double r);
 
 //-----------------------------------------------------------------------------
 
-QRadialGradient getBrushPad(int size, double hardness);
+QRadialGradient getBrushPad(double size, double hardness);
 
 //-----------------------------------------------------------------------------
